refactor: изменить тип id с int на long

+++ src/main/java/ru/yandex/practicum/filmorate/controllers/FilmController.java
@@ -22,8 +22,8 @@ import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
 @RequestMapping("/films")
 @Slf4j
 public class FilmController {
-    private static int id = 1;
-    private final Map<Integer, Film> films = new HashMap<>();
+    private static long id = 1;
+    private final Map<Long, Film> films = new HashMap<>();
 
     @PostMapping
     public Film addFilm(@Valid @RequestBody Film newFilm) {
