+++ src/main/java/ru/yandex/practicum/filmorate/controllers/UserController.java
@@ -22,8 +22,8 @@ import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
 @RequestMapping("/users")
 @Slf4j
 public class UserController {
-    private static int id = 1;
-    private final Map<Integer, User> users = new HashMap<>();
+    private static long id = 1;
+    private final Map<Long, User> users = new HashMap<>();
 
     @PostMapping
     public User addUser(@Valid @RequestBody User newUser) {
